+++ Desktop/ollamadev-ade/boson.config.php
@@ -1,4 +1,4 @@
 <?php
 
-define('OLLAMADEV_VERSION', '0.8.93');
+define('OLLAMADEV_VERSION', '0.8.94');
 define('OLLAMADEV_ADE', true);